ADMIN: Bring '-e .' into req'ts files
Now that the build requirements are specified in pyproject.toml,
including '-e .' in the requirements files doesn't lead to all
sorts of errory badness due to missing requirements.

There *may* be a better way to structure the project so that
it doesn't get imported on build... or, it might be that
an import of the project is inevitable, and this setup
is about as good as it can get.