DEV: Add types and triple modules
IOTriple combines MultiItemIterable and TupleContextManager,
and provides properties `stdin`, `stdout` and `stderr` for
literate referencing the three members of the tuple.

MultiItemIterable provides _map(), _all() and _any() for
iterables, with suppress versions able to ignore exceptions.

Base TupleContextManager creates a consistent MRO of tuple and
AbstractContextManager.

ClosingStdioTuple is a sample context manager which performs close()
on each item in the tuple.

TextIOTriple requires items to be a TextIOBase.

AutoIOTriple can be used to create either a TextIOTriple or a
FakeIOTriple, depending on whether the items are all a TextIOBase.
This will allow attaching different implementations to each.

AutoIOTuple is used to capture `sys.__std*__` and `sys.std*` as
they existed at module import.