DEV: Add types module
StdioTupleBase combines MultiItemTuple and TupleContextManager,
and provides properties `stdin`, `stdout` and `stderr` for
literate referencing the three members of the tuple.

MultiItemTuple provides _map(), _all() and _any() for iterables,
with suppress versions able to ignore exceptions.

Base TupleContextManager creates a consistent MRO of tuple and
AbstractContextManager.

ClosingStdioTuple is a sample context manager which performs close()
on each item in the tuple.

TextIOTuple requires items to be a TextIOBase.

AnyIOTuple can be used to create either a TextIOTuple or a FakeIOTuple,
depending on whether the items are all a TextIOBase.  This will allow
attaching different implementations to each.

AnyIOTuple is used to capture `sys.__std*__` and `sys.std*` as they
existed at module import.